# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoiça API",
        "description": "API de gestion de devis et factures pour artisans.",
        "contact": {
            "email": "contact@invoica.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Développement"
        },
        {
            "url": "http://my-default-host.com",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/v1/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Liste des documents",
                "operationId": "62975a8e151b8d8fc0b98b3e69be7def",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "quote",
                                "invoice"
                            ]
                        }
                    },
                    {
                        "name": "client_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des documents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Document"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Créer un document (devis/facture)",
                "operationId": "f11aae625102d887c912fe7a8cc678f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "quote",
                                        "enum": [
                                            "quote",
                                            "invoice"
                                        ]
                                    },
                                    "project_name": {
                                        "type": "string",
                                        "example": "Rénovation cuisine"
                                    },
                                    "issue_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-21"
                                    },
                                    "valid_until": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "subtotal": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "labor_cost": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "transport_cost": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "other_cost": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "total": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "total_in_words": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/documents/{document}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Afficher un document",
                "operationId": "0a86888d32d9a9165059255e7a8b70d3",
                "responses": {
                    "200": {
                        "description": "Détails du document avec articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Documents"
                ],
                "summary": "Modifier un document",
                "operationId": "00532d21041ac433edbef99c71a47c83",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project_name": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "total": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Supprimer un document",
                "operationId": "8215d99153c61fc0a495a9a9e437be25",
                "responses": {
                    "200": {
                        "description": "Document supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/documents/{document}/download": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Télécharger le PDF",
                "operationId": "a43a75f4c96bcb4a60a1189527190391",
                "responses": {
                    "200": {
                        "description": "PDF en base64",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "filename": {
                                            "type": "string"
                                        },
                                        "content": {
                                            "description": "PDF en base64",
                                            "type": "string"
                                        },
                                        "content_type": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/documents/{document}/email": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Envoyer le document par email",
                "operationId": "8d98d96bc11aa744344507f35b223df6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "client@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email envoyé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Company": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "logo": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string"
                    },
                    "manager_name": {
                        "type": "string"
                    },
                    "signature": {
                        "type": "string",
                        "nullable": true
                    },
                    "stamp": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Client": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Document": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "company_id": {
                        "type": "integer"
                    },
                    "client_id": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "quote",
                            "invoice"
                        ]
                    },
                    "number": {
                        "type": "string"
                    },
                    "project_name": {
                        "type": "string"
                    },
                    "issue_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "valid_until": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "subtotal": {
                        "type": "string"
                    },
                    "labor_cost": {
                        "type": "string"
                    },
                    "transport_cost": {
                        "type": "string"
                    },
                    "other_cost": {
                        "type": "string"
                    },
                    "total": {
                        "type": "string"
                    },
                    "total_in_words": {
                        "type": "string",
                        "nullable": true
                    },
                    "pdf_template": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "client": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Client"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    },
                    "company": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Company"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DocumentItem"
                        }
                    }
                },
                "type": "object"
            },
            "DocumentItem": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "document_id": {
                        "type": "integer"
                    },
                    "designation": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "string"
                    },
                    "unit_price": {
                        "type": "string"
                    },
                    "total_price": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Entrez votre token Sanctum (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}